backend/pyproject.toml
[project]
name = "alchemsim-backend"
version = "0.1.0"
description = "AlchemSim - SaaS CFD Simulation Platform Backend"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic[email]>=2.10",
    "pydantic-settings>=2.6",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.18",
    "boto3>=1.35",
    "celery[redis]>=5.4",
    "redis>=5.2",
    "gmsh>=4.13",
    "meshio>=5.3",
    "numpy>=2.1",
    "websockets>=14.1",
    "httpx>=0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "aiosqlite>=0.20",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
